refactor(test): 테스트 패키지 애플리케이션 패키지와 동리하게 변경
