+++ src/test/java/racingcar/domain/RacingCircuitTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
@@ -9,8 +9,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.MoveResults;
-import racingcar.domain.RacingCircuit;
 import racingcar.generator.RandomNumberFakeGenerator;
 
 @DisplayName("서킷 테스트")
